requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
abseil-cpp=20210324.2=hd77b12b_0
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py39h2bbff1b_1
aiosignal=1.2.0=pyhd3eb1b0_0
astetik=1.13=pypi_0
astor=0.8.1=py39haa95532_0
asttokens=2.0.5=pypi_0
astunparse=1.6.3=py_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py39haa95532_0
bottleneck=1.3.4=py39h080aedc_0
brotli=1.0.9=h8ffe710_7
brotli-bin=1.0.9=h8ffe710_7
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2022.5.18.1=h5b45459_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.5.18.1=py39hcbf5309_0
cffi=1.15.0=py39h2bbff1b_1
chances=0.1.9=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
clang=5.0=pypi_0
click=8.0.4=py39haa95532_0
colorama=0.4.4=pyhd3eb1b0_0
cryptography=3.4.8=py39h71e12ea_0
cudatoolkit=11.3.1=h59b6b97_2
cudnn=8.2.1=cuda11.3_0
cycler=0.11.0=pyhd8ed1ab_0
cython=0.29.28=py39hd77b12b_0
daal=2021.5.4=pypi_0
daal4py=2021.5.4=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
decorator=5.1.1=pypi_0
docker-pycreds=0.4.0=py_0
executing=0.8.3=pypi_0
filelock=3.7.0=pyhd8ed1ab_0
flatbuffers=1.12=pypi_0
fonttools=4.33.3=py39hb82d6ee_0
freetype=2.10.4=h546665d_1
frozenlist=1.2.0=py39h2bbff1b_0
future=0.18.2=py39hcbf5309_5
gast=0.4.0=pyhd3eb1b0_0
gensim=4.1.2=py39hd77b12b_0
geonamescache=1.3.0=pypi_0
giflib=5.2.1=h62dcd97_0
gitdb=4.0.9=pyhd8ed1ab_0
gitpython=3.1.27=pyhd8ed1ab_0
google-auth=1.35.0=pypi_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyhd3eb1b0_0
graphviz=2.38=hfd603c8_2
grpcio=1.42.0=py39hc60d5dd_0
h5py=3.6.0=py39h3de5c98_0
hdf5=1.10.6=h7ebc959_0
huggingface_hub=0.7.0=pyhd8ed1ab_0
icc_rt=2019.0.0=h0cc432a_1
icu=68.1=h6c2663c_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39haa95532_0
importlib_metadata=4.11.3=hd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipython=8.4.0=pypi_0
isort=5.9.3=pyhd3eb1b0_0
jedi=0.18.1=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
keras-base=2.6.0=pyhd3eb1b0_0
keras-gpu=2.6.0=hd3eb1b0_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kerasplotlib=1.0=pypi_0
kiwisolver=1.4.2=py39h2e07f2f_1
libbrotlicommon=1.0.9=h8ffe710_7
libbrotlidec=1.0.9=h8ffe710_7
libbrotlienc=1.0.9=h8ffe710_7
libclang=11.1.0=default_h5c34c98_1
libcurl=7.82.0=h86230a5_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.17.2=h23ce68f_1
libssh2=1.10.0=hcd4344a_0
libtiff=4.0.10=h21b02b4_1005
libuv=1.43.0=h8ffe710_0
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
lightgbm=3.3.2=pypi_0
lz4-c=1.9.3=h8ffe710_1
markdown=3.3.4=py39haa95532_0
matplotlib=3.5.2=py39hcbf5309_0
matplotlib-base=3.5.2=py39h581301d_0
matplotlib-inline=0.1.3=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
multidict=5.1.0=py39h2bbff1b_2
munkres=1.1.4=pyh9f0ad1d_0
ninja=1.11.0=h2d74725_0
numexpr=2.8.1=py39hb80d3ca_0
numpy=1.21.5=py39h7a0a035_2
numpy-base=1.21.5=py39hca35cd5_2
oauthlib=3.2.0=pyhd3eb1b0_0
openssl=1.1.1o=h8ffe710_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39hd77b12b_0
parso=0.8.3=pypi_0
pathtools=0.1.2=pypi_0
patsy=0.5.2=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
promise=2.3=py39hcbf5309_6
prompt-toolkit=3.0.29=pypi_0
protobuf=3.17.2=py39hd77b12b_0
psutil=5.9.1=py39hb82d6ee_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pybind11=2.9.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=pypi_0
pydotplus=2.0.2=py_3
pygments=2.12.0=pypi_0
pyjwt=2.1.0=py39haa95532_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.12.3=py39hb0d2dfa_4
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyreadline=2.1=py39haa95532_1
pysocks=1.7.1=py39haa95532_0
pythainlp=3.0.8=pypi_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-graphviz=0.20=pypi_0
python_abi=3.9=2_cp39
pytorch=1.10.2=cpu_py39h907fbb5_0
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py39h2bbff1b_1
qt=5.12.9=h5909a2a_4
regex=2022.4.24=py39hb82d6ee_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
sacremoses=0.0.53=pyhd8ed1ab_0
scikit-learn=1.0.2=py39hf11a4ad_1
scikit-learn-intelex=2021.5.4=pypi_0
scipy=1.7.3=py39h0a974cb_0
seaborn=0.11.2=pypi_0
sentry-sdk=1.5.12=pyhd8ed1ab_0
setproctitle=1.2.3=py39hb82d6ee_0
setuptools=61.2.0=py39haa95532_0
shortuuid=1.0.9=pypi_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smart_open=5.2.1=py39haa95532_0
smmap=5.0.0=pypi_0
snappy=1.1.9=h6c2663c_0
sqlite=3.38.2=h2bbff1b_0
stack-data=0.2.0=pypi_0
statsmodels=0.13.2=pypi_0
talos=1.3=pypi_0
tbb=2021.6.0=pypi_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py39haa95532_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.6.0=gpu_py39he88c5ba_0
tensorflow-addons=0.13.0=py39_2
tensorflow-base=2.6.0=gpu_py39hb3da07e_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
tensorflow-gpu=2.6.0=h17022bd_0
termcolor=1.1.0=py39haa95532_1
threadpoolctl=2.2.0=pyh0d69192_0
tinydb=4.7.0=pypi_0
tk=8.6.12=h8ffe710_0
tokenizers=0.12.1=py39h8c9392a_1
tornado=6.1=py39hb82d6ee_3
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.2.2.post1=pypi_0
transformers=4.19.2=pyhd8ed1ab_0
typeguard=2.13.0=py_1
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
unicodedata2=14.0.0=py39hb82d6ee_1
urllib3=1.26.9=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.12.18=pyhd8ed1ab_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.35.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
wrangle=0.7.2=pypi_0
wrapt=1.13.3=py39h2bbff1b_2
xgboost=1.6.1=pypi_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yarl=1.6.3=py39h2bbff1b_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.9=h6255e5f_0
